Select a random word for a game
Created a module called Selectable that has a few methods to select
words that are between 5 and 12 characters long; one of these words
will be used in a hangman game.

The load_words method just opens the file that contains all the english
words and load them in a variable; with the trim_words method it removes
the new line character found when loading the words on the array; the
filter_words method removes words that are less than 5 characters long.
Finally, set_dictionary method uses the previous methods to form the
dictionary used to get one word by the Game class.

The Game class selects a word each time an instance is created. It
contains the Selectable module to do this action and using the sample
method it selects one word and store it as a class variable.

At the end of the script there is an instance to check if it works
correctly.