Show word without letters
The update_secret_word method uses the game's word and through a loop it
adds '_' to a class variable that will be updated when the player
guesses a letter. And the display secret_word method just shows this
class variable to let the player know his progress in the game.