Add more methods to structure a turn
Fixed a bug in obtain_player_letter method that showed an extra message
when the player entered a valid letter; add method to remove a player
try when he enters a letter that is not part of the word. And finally,
added a method that adds one when a turn has finished.