Fix bug on guessed letters
On a game, when the player guessed a letter, there was a bug that
allowed the player to keep entering the guessed letters and the turns
were advancing; in a traditional game that behaviour is not allowed.

Now when the player enters a letter, it also checks with the
guessed_letters variable that the letter is not included.